OV-31: + move timeaxisarkers to helpers\time-axis-markers.ts
